### 2020-08-14 #2574 @whedon
[PRE REVIEW]: Foolbox Native: Fast adversarial attacks to benchmark the robustness of machine learning models in PyTorch, TensorFlow, and JAX
**Submitting author:** @jonasrauber (<a href="http://orcid.org/0000-0001-6795-9441">Jonas Rauber</a>)
**Repository:** <a href="https://github.com/bethgelab/foolbox" target ="_blank">https://github.com/bethgelab/foolbox</a>
**Version:** v3.1.0
