### 2020-08-15 #2574 @jonasrauber
We had one small technical issue when creating the paper: the „joint senior authors“ footnote could only be added to one senior author. This seems to be a limitation of the Markdown to LaTeX conversion.
Any ideas how we can fix this?
How do you usually indicate joint first authors?
