### 2020-08-16 e149a5e
Make HSJA more robust to numerical imprecision
