### 2020-08-16 4dbd575
Change data_format implementation
