### 2020-08-17 #2574 @kyleniemeyer
> We had one small technical issue when creating the paper: the „joint senior authors“ footnote could only be added to one senior author. This seems to be a limitation of the Markdown to LaTeX conversion.
> Any ideas how we can fix this?
> How do you usually indicate joint first authors?

@openjournals/dev any ideas about this author footnote issue?
