### 2020-08-17 c4b526b
Fix indexing to avoid error on JAX
