### 2020-08-17 ed8fbba
Cleanup code
