### 2020-08-21 #2574 @arfon
> @openjournals/dev any ideas about this author footnote issue?

The way I recommend doing this is with a footnote next to the name of the first (or second) author stating that they are joint first authors e.g.:

```
authors:
  - name: First Author
    orcid: 0000-0003-0872-7098
    affiliation: "1, 2"
  - name: Second Author^[Joint first author]
    affiliation: 2
```
