### 2020-08-24 #2574 @jonasrauber
@terrytangyuan

Here are some suggestions for potential reviewers from the list:

- ethanwharris
- rasbt
- willprice

And some more potential reviewers from the list:

- GregaVrbancic
- JesseLivezey
- nilesh-patil

It’s difficult to find someone from the specific field (adversarial robustness) on the list, so here is another suggestion, but he is not (yet) on the list:

- ttbrunner (Note: we are co-authors on a book chapter for a competition that we organized and in which ttbrunner participated; I don’t think it’s an actual COI, but it may be perceived as one)
