### 2020-08-28 #2574 @GregaVrbancic
Hi @terrytangyuan, I am available to review this submission, but unfortunately, not before the end of next week. Sadly, I do not work in this specific field but I do most of my work training deep CNNs, so I think I would be able to adequately review this submission.
