### 2020-08-29 5f0e12f
Implement HopSkipJump attack (#574)
* Add frist draft of HSJ attack

* Finalize HSJ attack

* Fix types

* Fix gradient estimation

* Make unit tests faster and require less memory

* Improve stability of attack

* Increase test coverage

* Increase stability of HSJ attack and tests

* Remove debugging statements

* Add comment

* Fix typo

* Increase stability of HSJ attack and tests

* Fix typo

* Make HSJA more robust to numerical imprecision

* Fix test timeout

* Make tests faster
