+++ foolbox/attacks/hop_skip_jump.py
@@ -1,0 +1,366 @@
+import logging
+from typing import Union, Any, Optional, Callable, List
+from typing_extensions import Literal
+
+import math
+
+import eagerpy as ep
+import numpy as np
+
+from foolbox.attacks import LinearSearchBlendedUniformNoiseAttack
+from foolbox.tensorboard import TensorBoard
+from ..models import Model
+
+from ..criteria import Criterion
+
+from ..distances import l1
+
+from ..devutils import atleast_kd, flatten
+
+from .base import MinimizationAttack, get_is_adversarial
+from .base import get_criterion
+from .base import T
+from .base import raise_if_kwargs
+from ..distances import l2, linf
+
+
+class HopSkipJump(MinimizationAttack):
+    """A powerful adversarial attack that requires neither gradients
+    nor probabilities [#Chen19].
+
+    Args:
+        init_attack : Attack to use to find a starting points. Defaults to
+            LinearSearchBlendedUniformNoiseAttack. Only used if starting_points is None.
+        steps : Number of optimization steps within each binary search step.
+        initial_gradient_eval_steps: Initial number of evaluations for gradient estimation.
+            Larger initial_num_evals increases time efficiency, but
+            may decrease query efficiency.
+        max_gradient_eval_steps : Maximum number of evaluations for gradient estimation.
+        stepsize_search : How to search for stepsize; choices are 'geometric_progression',
+            'grid_search'. 'geometric progression' initializes the stepsize
+            by ||x_t - x||_p / sqrt(iteration), and keep decreasing by half
+            until reaching the target side of the boundary. 'grid_search'
+            chooses the optimal epsilon over a grid, in the scale of
+            ||x_t - x||_p.
+        gamma : The binary search threshold theta is gamma / d^1.5 for
+                   l2 attack and gamma / d^2 for linf attack.
+        tensorboard : The log directory for TensorBoard summaries. If False, TensorBoard
+            summaries will be disabled (default). If None, the logdir will be
+            runs/CURRENT_DATETIME_HOSTNAME.
+        constraint : Norm to minimize, either "l2" or "linf"
+
+    References:
+        .. [#Chen19] Jianbo Chen, Michael I. Jordan, Martin J. Wainwright,
+        "HopSkipJumpAttack: A Query-Efficient Decision-Based Attack",
+        https://arxiv.org/abs/1904.02144
+    """
+
+    distance = l1
+
+    def __init__(
+        self,
+        init_attack: Optional[MinimizationAttack] = None,
+        steps: int = 64,
+        initial_gradient_eval_steps: int = 100,
+        max_gradient_eval_steps: int = 10000,
+        stepsize_search: Union[
+            Literal["geometric_progression"], Literal["grid_search"]
+        ] = "geometric_progression",
+        gamma: float = 1.0,
+        tensorboard: Union[Literal[False], None, str] = False,
+        constraint: Union[Literal["linf"], Literal["l2"]] = "l2",
+    ):
+        if init_attack is not None and not isinstance(init_attack, MinimizationAttack):
+            raise NotImplementedError
+        self.init_attack = init_attack
+        self.steps = steps
+        self.initial_num_evals = initial_gradient_eval_steps
+        self.max_num_evals = max_gradient_eval_steps
+        self.stepsize_search = stepsize_search
+        self.gamma = gamma
+        self.tensorboard = tensorboard
+        self.constraint = constraint
+
+        assert constraint in ("l2", "linf")
+        if constraint == "l2":
+            self.distance = l2
+        else:
+            self.distance = linf
+
+    def run(
+        self,
+        model: Model,
+        inputs: T,
+        criterion: Union[Criterion, T],
+        *,
+        early_stop: Optional[float] = None,
[…]
