### 2020-09-12 bethgelab/foolbox#581 @ethanwharris
[JOSS Review] State of the field
Hi, I'm reviewing the Foolbox native paper on JOSS ([openjournals/joss-reviews/issues/2607](https://github.com/openjournals/joss-reviews/issues/2607)). In order to satisfy the [state of the field check in the software paper review criteria](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper), it would be good if you could include some discussion of other libraries that are focussed on adversarial robustness research. The two other commonly-used packages that I know of (may not be an exhaustive list) and would be good to briefly compare to are [CleverHans](https://github.com/tensorflow/cleverhans) and [AdverTorch](https://github.com/BorealisAI/advertorch).
