### 2020-09-15 2d978fc
Simplfy structure used to test attacks
