+++ tests/test_attacks_raise.py
@@ -3,8 +3,61 @@ import pytest
 import eagerpy as ep
 import foolbox as fbn
 
+from conftest import ModelDescriptionAndData
+
 L2 = fbn.types.L2
 Linf = fbn.types.Linf
+
+
+def test_singlepixel_run_raises() -> None:
+    class Model:
+        def __call__(self, inputs: Any) -> Any:
+            return inputs.mean(axis=(2, 3))
+
+    model = Model()
+    with pytest.raises(ValueError):
+        fbn.NumPyModel(model, bounds=(0, 1), data_format="foo")
+
+    fmodel = fbn.NumPyModel(model, bounds=(0, 1))
+    x, y = ep.astensors(
+        *fbn.samples(
+            fmodel, dataset="imagenet", batchsize=16, data_format="channels_first"
+        )
+    )
+
+    with pytest.raises(ValueError, match="channel_axis"):
+        fbn.attacks.SinglePixelAttack().run(fmodel, x, y)
+
+    with pytest.raises(ValueError, match="channel_axis"):
+        fbn.attacks.SinglePixelAttack(channel_axis=2).run(fmodel, x, y)
+
+
+def test_localsearch_run_raises() -> None:
+    class Model:
+        def __call__(self, inputs: Any) -> Any:
+            return inputs.mean(axis=(2, 3))
+
+    model = Model()
+    with pytest.raises(ValueError):
+        fbn.NumPyModel(model, bounds=(0, 1), data_format="foo")
+
+    fmodel = fbn.NumPyModel(model, bounds=(0, 1))
+    x, y = ep.astensors(
+        *fbn.samples(
+            fmodel, dataset="imagenet", batchsize=16, data_format="channels_first"
+        )
+    )
+
+    with pytest.raises(ValueError, match="channel_axis"):
+        fbn.attacks.LocalSearchAttack().run(fmodel, x, fbn.TargetedMisclassification(y))
+
+    with pytest.raises(ValueError, match="channel_axis"):
+        fbn.attacks.LocalSearchAttack(channel_axis=2).run(
+            fmodel, x, fbn.TargetedMisclassification(y)
+        )
+
+    with pytest.raises(ValueError, match="unsupported criterion"):
+        fbn.attacks.LocalSearchAttack().run(fmodel, x, y)
 
 
 def test_ead_init_raises() -> None:
@@ -40,7 +93,7 @@ def test_genattack_numpy(request: Any) -> None:
 
 
 def test_deepfool_run_raises(
-    fmodel_and_data_ext_for_attacks: Tuple[Tuple[fbn.Model, ep.Tensor, ep.Tensor], bool]
+    fmodel_and_data_ext_for_attacks: ModelDescriptionAndData,
 ) -> None:
     (fmodel, x, y), _ = fmodel_and_data_ext_for_attacks
     if isinstance(x, ep.NumPyTensor):
@@ -52,7 +105,7 @@ def test_deepfool_run_raises(
 
 
 def test_blended_noise_attack_run_warns(
-    fmodel_and_data_ext_for_attacks: Tuple[Tuple[fbn.Model, ep.Tensor, ep.Tensor], bool]
+    fmodel_and_data_ext_for_attacks: ModelDescriptionAndData,
 ) -> None:
     (fmodel, x, y), _ = fmodel_and_data_ext_for_attacks
     attack = fbn.attacks.LinearSearchBlendedUniformNoiseAttack(directions=1)
[…]
