+++ tests/conftest.py
@@ -1,14 +1,19 @@
-from typing import Optional, Callable, Tuple, Dict, Any, List
+from typing import Optional, Callable, Tuple, Dict, Any, List, NamedTuple
 import functools
 import pytest
 import eagerpy as ep
 
-import foolbox
-import foolbox as fbn
+import foolbox as fb
 
-ModelAndData = Tuple[fbn.Model, ep.Tensor, ep.Tensor]
+ModelAndData = Tuple[fb.Model, ep.Tensor, ep.Tensor]
+CallableModelAndDescription = NamedTuple(
+    "ModelDescription", [("model_fn", Callable[..., ModelAndData]), ("real", bool)],
+)
+ModelDescriptionAndData = NamedTuple(
+    "ModelDescriptionAndData", [("model_and_data", ModelAndData), ("real", bool)],
+)
 
-models: Dict[str, Tuple[Callable[..., ModelAndData], bool]] = {}
+models: Dict[str, CallableModelAndDescription] = {}
 models_for_attacks: List[str] = []
 
 
@@ -27,7 +32,7 @@ def dummy(request: Any) -> ep.Tensor:
 
 
 def register(
-    backend: str, *, real: bool, attack: bool = True
+    backend: str, *, real: bool = False, attack: bool = True
 ) -> Callable[[Callable], Callable]:
     def decorator(f: Callable[[Any], ModelAndData]) -> Callable[[Any], ModelAndData]:
         @functools.wraps(f)
@@ -39,7 +44,7 @@ def register(
         global models
         global real_models
 
-        models[model.__name__] = (model, real)
+        models[model.__name__] = CallableModelAndDescription(model_fn=model, real=real)
         if attack:
             models_for_attacks.append(model.__name__)
         return model
@@ -48,7 +53,7 @@ def register(
 
 
 def pytorch_simple_model(
-    device: Any = None, preprocessing: fbn.types.Preprocessing = None
+    device: Any = None, preprocessing: fb.types.Preprocessing = None
 ) -> ModelAndData:
     import torch
 
@@ -60,27 +65,27 @@ def pytorch_simple_model(
 
     model = Model().eval()
     bounds = (0, 1)
-    fmodel = fbn.PyTorchModel(
+    fmodel = fb.PyTorchModel(
         model, bounds=bounds, device=device, preprocessing=preprocessing
     )
 
-    x, _ = fbn.samples(fmodel, dataset="imagenet", batchsize=16)
+    x, _ = fb.samples(fmodel, dataset="cifar10", batchsize=16)
     x = ep.astensor(x)
     y = fmodel(x).argmax(axis=-1)
     return fmodel, x, y
 
 
-@register("pytorch", real=False)
+@register("pytorch")
 def pytorch_simple_model_default(request: Any) -> ModelAndData:
     return pytorch_simple_model()
 
 
-@register("pytorch", real=False)
+@register("pytorch")
 def pytorch_simple_model_default_flip(request: Any) -> ModelAndData:
     return pytorch_simple_model(preprocessing=dict(flip_axis=-3))
 
 
-@register("pytorch", real=False, attack=False)
+@register("pytorch", attack=False)
 def pytorch_simple_model_default_cpu_native_tensor(request: Any) -> ModelAndData:
     import torch
 
@@ -89,19 +94,19 @@ def pytorch_simple_model_default_cpu_native_tensor(request: Any) -> ModelAndData
     return pytorch_simple_model("cpu", preprocessing=dict(mean=mean, std=std, axis=-3))
 
 
-@register("pytorch", real=False, attack=False)
[…]
