+++ tests/test_binarization_attack.py
@@ -1,16 +1,16 @@
-from typing import Tuple
 import pytest
-import eagerpy as ep
 
-from foolbox import accuracy, Model
+from foolbox import accuracy
 from foolbox.models import ThresholdingWrapper
 from foolbox.devutils import flatten
 from foolbox.attacks import BinarySearchContrastReductionAttack
 from foolbox.attacks import BinarizationRefinementAttack
 
+from conftest import ModelDescriptionAndData
+
 
 def test_binarization_attack(
-    fmodel_and_data_ext_for_attacks: Tuple[Tuple[Model, ep.Tensor, ep.Tensor], bool],
+    fmodel_and_data_ext_for_attacks: ModelDescriptionAndData,
 ) -> None:
 
     # get a model with thresholding
