+++ tests/conftest.py
@@ -3,9 +3,10 @@ import functools
 import pytest
 import eagerpy as ep
 
-import foolbox as fb
+import foolbox
+import foolbox as fbn
 
-ModelAndData = Tuple[fb.Model, ep.Tensor, ep.Tensor]
+ModelAndData = Tuple[fbn.Model, ep.Tensor, ep.Tensor]
 CallableModelAndDescription = NamedTuple(
     "ModelDescription", [("model_fn", Callable[..., ModelAndData]), ("real", bool)],
 )
@@ -53,7 +54,7 @@ def register(
 
 
 def pytorch_simple_model(
-    device: Any = None, preprocessing: fb.types.Preprocessing = None
+    device: Any = None, preprocessing: fbn.types.Preprocessing = None
 ) -> ModelAndData:
     import torch
 
@@ -65,11 +66,11 @@ def pytorch_simple_model(
 
     model = Model().eval()
     bounds = (0, 1)
-    fmodel = fb.PyTorchModel(
+    fmodel = fbn.PyTorchModel(
         model, bounds=bounds, device=device, preprocessing=preprocessing
     )
 
-    x, _ = fb.samples(fmodel, dataset="cifar10", batchsize=16)
+    x, _ = fbn.samples(fmodel, dataset="imagenet", batchsize=16)
     x = ep.astensor(x)
     y = fmodel(x).argmax(axis=-1)
     return fmodel, x, y
@@ -115,10 +116,10 @@ def pytorch_simple_model_object(request: Any) -> ModelAndData:
 
 @register("pytorch", real=True)
 def pytorch_mnist(request: Any) -> ModelAndData:
-    fmodel = fb.zoo.ModelLoader.get().load(
+    fmodel = fbn.zoo.ModelLoader.get().load(
         "examples/zoo/mnist/", module_name="foolbox_model"
     )
-    x, y = fb.samples(fmodel, dataset="mnist", batchsize=16)
+    x, y = fbn.samples(fmodel, dataset="mnist", batchsize=16)
     x = ep.astensor(x)
     y = ep.astensor(y)
     return fmodel, x, y
@@ -133,16 +134,16 @@ def pytorch_resnet18(request: Any) -> ModelAndData:
 
     model = models.resnet18(pretrained=True).eval()
     preprocessing = dict(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225], axis=-3)
-    fmodel = fb.PyTorchModel(model, bounds=(0, 1), preprocessing=preprocessing)
+    fmodel = fbn.PyTorchModel(model, bounds=(0, 1), preprocessing=preprocessing)
 
-    x, y = fb.samples(fmodel, dataset="imagenet", batchsize=16)
+    x, y = fbn.samples(fmodel, dataset="imagenet", batchsize=16)
     x = ep.astensor(x)
     y = ep.astensor(y)
     return fmodel, x, y
 
 
 def tensorflow_simple_sequential(
-    device: Optional[str] = None, preprocessing: fb.types.Preprocessing = None
+    device: Optional[str] = None, preprocessing: fbn.types.Preprocessing = None
 ) -> ModelAndData:
     import tensorflow as tf
 
@@ -150,11 +151,11 @@ def tensorflow_simple_sequential(
         model = tf.keras.Sequential()
         model.add(tf.keras.layers.GlobalAveragePooling2D())
     bounds = (0, 1)
-    fmodel = fb.TensorFlowModel(
+    fmodel = fbn.TensorFlowModel(
         model, bounds=bounds, device=device, preprocessing=preprocessing
     )
 
[…]
