### 2020-09-15 c4f489a
Add targeted mode of PGD
