+++ foolbox/attacks/gradient_descent_base.py
@@ -9,7 +9,7 @@ from ..types import Bounds
 
 from ..models.base import Model
 
-from ..criteria import Misclassification
+from ..criteria import Misclassification, TargetedMisclassification
 
 from ..distances import l1, l2, linf
 
@@ -55,7 +55,7 @@ class BaseGradientDescent(FixedEpsilonAttack, ABC):
         self,
         model: Model,
         inputs: T,
-        criterion: Union[Misclassification, T],
+        criterion: Union[Misclassification, TargetedMisclassification, T],
         *,
         epsilon: float,
         **kwargs: Any,
@@ -65,18 +65,22 @@ class BaseGradientDescent(FixedEpsilonAttack, ABC):
         criterion_ = get_criterion(criterion)
         del inputs, criterion, kwargs
 
-        if not isinstance(criterion_, Misclassification):
+        # perform a gradient ascent (targeted attack) or descent (untargeted attack)
+        if isinstance(criterion_, Misclassification):
+            gradient_step_sign = 1.0
+            classes = criterion_.labels
+        elif isinstance(criterion_, TargetedMisclassification):
+            gradient_step_sign = -1.0
+            classes = criterion_.target_classes
+        else:
             raise ValueError("unsupported criterion")
 
-        labels = criterion_.labels
-        loss_fn = self.get_loss_fn(model, labels)
+        loss_fn = self.get_loss_fn(model, classes)
 
         if self.abs_stepsize is None:
             stepsize = self.rel_stepsize * epsilon
         else:
             stepsize = self.abs_stepsize
-
-        x = x0
 
         if self.random_start:
             x = self.get_random_start(x0, epsilon)
@@ -87,7 +91,7 @@ class BaseGradientDescent(FixedEpsilonAttack, ABC):
         for _ in range(self.steps):
             _, gradients = self.value_and_grad(loss_fn, x)
             gradients = self.normalize(gradients, x=x, bounds=model.bounds)
-            x = x + stepsize * gradients
+            x = x + gradient_step_sign * stepsize * gradients
             x = self.project(x, x0, epsilon)
             x = ep.clip(x, *model.bounds)
 
