### 2020-09-16 bethgelab/foolbox#581 @jonasrauber
Hi @ethanwharris, thanks for taking the time to review Foolbox Native. I am currently on holidays and will get back to you with a more detailed  answer next week. The main difference is that we support multiple frameworks, not just either PyTorch (AdverTorch) or TensorFlow (Cleverhans).
