### 2020-09-17 bethgelab/foolbox#585 @GregaVrbancic
[JOSS Review] References
Hello @jonasrauber, I am the reviewer of your Foolbox Native paper on JOSS (openjournals/joss-reviews/issues/2607). Please, when you are available fix the missing DOIs in your paper.

```
Reference check summary:

OK DOIs

- None

MISSING DOIs

- https://doi.org/10.1109/sp40000.2020.00045 may be missing for title: HopSkipJumpAttack: A query-efficient decision-based attack

INVALID DOIs

- None
```

_Originally posted by @whedon in https://github.com/openjournals/joss-reviews/issues/2607#issuecomment-682839171_
