### 2020-09-17 bethgelab/foolbox#586 @GregaVrbancic
[JOSS Review] Example usage
Hi @jonasrauber, I went through [the example using tensorflow](https://github.com/bethgelab/foolbox/blob/master/examples/tensorflow_resnet50.py) and while I did not have any issues running the code, I think it would be beneficial if the provided examples would be more clearly named, so the users could find the right example more easily. Additionally, I would also appreciate a few sentences in the examples on the topic of how to interpret the output results. This way the tool would be also more attractive to potential users who are not so familiar with the topic.

I also noticed, that there is a tutorial in a form of a jupyter notebook. Maybe it could be a bit more emphasized in the readme page so it can be found more easily, and also I think that the notebook tutorial would benefit a lot if you would improve it by adding a bit of results interpretation.
