### 2020-09-17 bethgelab/foolbox#587 @GregaVrbancic
[JOSS Review] Performance
Hi @jonasrauber, I summary of the paper and also in the readme is mentioned that the Foolbox was rewritten from scratch to achieve native performance. It would be interesting if you could provide us with some performance comparison (maybe in a form of chart or table) between the current version and the previous one.
