### 2020-09-17 #2607 @GregaVrbancic
Hello, @terrytangyuan and @jonasrauber, I have completed my review and opened a few issues at the Foolbox repository:

- [Example usage](https://github.com/bethgelab/foolbox/issues/586)
- [Performance](https://github.com/bethgelab/foolbox/issues/587)
- [References](https://github.com/bethgelab/foolbox/issues/585)

@terrytangyuan, additionally, I have noticed that one of the authors (Matthias Bethge) is not the contributor of the project. I am not sure if all of the authors needs to be significant contributors to the project or not. If not, please ignore this remark.
