### 2020-09-17 #2607 @terrytangyuan
Thanks. @jonasrauber Could you look into those issues? Also please clarify Matthias Bethge's contributions if not observable from commit history (which is usual sometimes).
