### 2020-09-21 98c1121
Fix assertion
