+++ paper/paper.bib
@@ -92,6 +92,8 @@
   booktitle={2020 IEEE Symposium on Security and Privacy (SP)},
   pages={1277--1294},
   year={2020},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/SP40000.2020.00045},
+  url={http://dx.doi.org/10.1109/SP40000.2020.00045}
 }
 
