### 2020-09-22 bethgelab/foolbox#585 @jonasrauber
Fixed in https://github.com/bethgelab/foolbox/commit/26462d95a769036e4ac4aca8bab67381f200d645

Checked by whedon: https://github.com/openjournals/joss-reviews/issues/2607#issuecomment-696604651
