### 2020-09-22 6e4c02e
added a performance README
