### 2020-09-22 3b3d6f3
added performance section to README
