### 2020-09-22 bethgelab/foolbox#587 @jonasrauber
Good point @GregaVrbancic. I have now added a simple performance comparison to the repository: https://github.com/bethgelab/foolbox/tree/master/performance
