### 2020-09-23 0012148
updated the spatial_attack_pytorch_resnet18 example
