### 2020-09-23 1c9e529
updated the substituion_model_pytorch_resnet18 example
