+++ examples/substituion_model_pytorch_resnet18.py
@@ -1,5 +1,11 @@
 #!/usr/bin/env python3
 # mypy: no-disallow-untyped-defs
+"""
+Sometimes one wants to replace the gradient of a model with a different gradient
+from another model to make the attack more reliable. That is, the forward pass
+should go through model 1, but the backward pass should go through model 2.
+This example shows how that can be done in Foolbox.
+"""
 import torchvision.models as models
 import eagerpy as ep
 from foolbox import PyTorchModel, accuracy, samples
@@ -7,8 +13,8 @@ from foolbox.attacks import LinfPGD
 from foolbox.attacks.base import get_criterion
 
 
-if __name__ == "__main__":
-    # instantiate a model
+def main() -> None:
+    # instantiate a model (could also be a TensorFlow or JAX model)
     model = models.resnet18(pretrained=True).eval()
     preprocessing = dict(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225], axis=-3)
     fmodel = PyTorchModel(model, bounds=(0, 1), preprocessing=preprocessing)
@@ -17,7 +23,8 @@ if __name__ == "__main__":
     # wrapping the tensors with ep.astensors is optional, but it allows
     # us to work with EagerPy tensors in the following
     images, labels = ep.astensors(*samples(fmodel, dataset="imagenet", batchsize=16))
-    print(accuracy(fmodel, images, labels))
+    clean_acc = accuracy(fmodel, images, labels) * 100
+    print(f"clean accuracy:  {clean_acc:.1f} %")
 
     # replace the gradient with the gradient from another model
     model2 = fmodel  # demo, we just use the same model
@@ -41,14 +48,27 @@ if __name__ == "__main__":
     epsilons = [0.0, 0.001, 0.01, 0.03, 0.1, 0.3, 0.5, 1.0]
     advs, _, success = attack(fmodel, images, labels, epsilons=epsilons)
 
-    # calculate and report the robust accuracy
+    # calculate and report the robust accuracy (the accuracy of the model when
+    # it is attacked)
     robust_accuracy = 1 - success.float32().mean(axis=-1)
+    print("robust accuracy for perturbations with")
     for eps, acc in zip(epsilons, robust_accuracy):
-        print(eps, acc.item())
+        print(f"  Linf norm ≤ {eps:<6}: {acc.item() * 100:4.1f} %")
 
     # we can also manually check this
+    print()
+    print("manual check:")
+    print()
+    print("robust accuracy for perturbations with")
     for eps, advs_ in zip(epsilons, advs):
-        print(eps, accuracy(fmodel, advs_, labels))
+        acc2 = accuracy(fmodel, advs_, labels)
+        print(f"  Linf norm ≤ {eps:<6}: {acc2 * 100:4.1f} %")
         # but then we also need to look at the perturbation sizes
         # and check if they are smaller than eps
-        print((advs_ - images).norms.linf(axis=(1, 2, 3)).numpy())
+        print(f"    if the perturbation sizes are all smaller then {eps}")
+        perturbation_sizes = (advs_ - images).norms.linf(axis=(1, 2, 3)).numpy()
+        print("    ", str(perturbation_sizes).replace("\n", "\n" + "    "))
+
+
+if __name__ == "__main__":
+    main()
