### 2020-09-23 1bd50eb
use clipped advs in substituion_model_pytorch_resnet18 example
