### 2020-09-23 532a498
improved the examples, added comments, and more detailed output strings
