+++ examples/multiple_attacks_pytorch_resnet18.py
@@ -7,7 +7,7 @@ import numpy as np
 
 
 if __name__ == "__main__":
-    # instantiate a model
+    # instantiate a model (could also be a TensorFlow or JAX model)
     model = models.resnet18(pretrained=True).eval()
     preprocessing = dict(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225], axis=-3)
     fmodel = PyTorchModel(model, bounds=(0, 1), preprocessing=preprocessing)
@@ -16,8 +16,8 @@ if __name__ == "__main__":
     # wrapping the tensors with ep.astensors is optional, but it allows
     # us to work with EagerPy tensors in the following
     images, labels = ep.astensors(*samples(fmodel, dataset="imagenet", batchsize=16))
-    print("accuracy")
-    print(accuracy(fmodel, images, labels))
+    clean_acc = accuracy(fmodel, images, labels)
+    print(f"clean accuracy:  {clean_acc * 100:.1f} %")
     print("")
 
     attacks = [
@@ -58,9 +58,16 @@ if __name__ == "__main__":
         print(attack)
         print("  ", 1.0 - success_.mean(axis=-1).round(2))
 
+    # calculate and report the robust accuracy (the accuracy of the model when
+    # it is attacked) using the best attack per sample
     robust_accuracy = 1.0 - attack_success.max(axis=0).mean(axis=-1)
     print("")
     print("-" * 79)
     print("")
     print("worst case (best attack per-sample)")
     print("  ", robust_accuracy.round(2))
+    print("")
+
+    print("robust accuracy for perturbations with")
+    for eps, acc in zip(epsilons, robust_accuracy):
+        print(f"  Linf norm ≤ {eps:<6}: {acc.item() * 100:4.1f} %")
