### 2020-09-23 df5985e
added open in colab image
