### 2020-09-23 bethgelab/foolbox#586 @jonasrauber
Thanks for the feedback.
I have updated the examples to contain more comments and explanations, renamed them to be more consistent and explicit, and added a README to the examples folder (https://github.com/bethgelab/foolbox/pull/590).
Furthermore, I have updated the Jupyter notebook to include some explanations of the results and I added a direct link to open the notebook in Google Colab to the the main README (that gives it additional visibility, because the link is an image).
