### 2020-09-23 8b50025
added comparison to other libraries
