+++ paper/paper.md
@@ -41,6 +41,8 @@ Evaluating the adversarial robustness of machine learning models is crucial to u
 
 This is great for both users and developers of adversarial attacks. Users can efficiently evaluate the robustness of different models in different frameworks using the same set of state-of-the-art adversarial attacks, thus obtaining comparable results. Attack developers do not need to choose between supporting just one framework or reimplementing their new adversarial attack multiple times and dealing with code duplication. In addition, they both benefit from the comprehensive type annotations [@pep484] in Foolbox Native to catch bugs even before running their code.
 
+The combination of being framework-agnostic and simultaneously achieving native performance sets Foolbox Native apart from other adversarial attack libraries. The most popular alternative to Foolbox is CleverHans^[https://github.com/tensorflow/cleverhans]. It was the first adversarial attack library and has traditionally focused solely on TensorFlow (plans to make it framework-agnostic *in the future* have been announced). The original Foolbox was the second adversarial attack library and the first one to be framework-agnostic. Back then, this was achieved at the expense of performance. The adversarial robustness toolbox ART^[https://github.com/Trusted-AI/adversarial-robustness-toolbox] is another framework-agnostic adversarial attack library, but it is conceptually inspired by the original Foolbox and thus comes with the same performance trade-off. AdverTorch^[https://github.com/BorealisAI/advertorch] is another popular adversarial attack library that was inspired by the original Foolbox but improved its performance by focusing soley on PyTorch. Foolbox Native is our attempt to improve the performance without sacrificing the framework-agnostic design that is crucial to consistently evaluate the robustness of different machine learning models that use different frameworks.
+
 # Use Cases
 
 Foolbox was designed to make adversarial attacks easy to apply even without expert knowledge. It has been used in numerous scientific publications and has already been cited more than 220 times. On GitHub it has received contributions from several developers and has gathered more than 1.500 stars. It provides the reference implementations of various adversarial attacks, including the Boundary Attack [@brendel2018decisionbased], the Pointwise Attack [@schott2018towards], clipping-aware noise attacks [@rauber2020fast], the Brendel Bethge Attack [@brendel2019accurate], and the HopSkipJump Attack [@chen2020hopskipjumpattack], and is under active development since 2017.
