### 2020-09-23 bethgelab/foolbox#581 @jonasrauber
I added a paragraph comparing Foolbox Native to CleverHans, AdverTorch, and ART (https://github.com/bethgelab/foolbox/commit/8b500251f4b741da09a6959c39aa8af1ce038f13) that should make the difference clear.
