### 2020-09-23 #2607 @jonasrauber
@GregaVrbancic @ethanwharris Thanks for taking the time to review Foolbox Native. I have addressed the four issues you opened. Thanks for the helpful feedback. Let me know if there is anything else you think I should change.

@terrytangyuan Matthias Bethge provided high-level supervision and guidance for the project and helped editing the paper.
