### 2020-09-23 82cf0c8
Improve type annotations
