### 2020-09-23 b112bef
Improve type annotations
