### 2020-09-23 8d814fe
Remove unused type ignore comment
