### 2020-09-23 3f8467b
Fix type annotation
