### 2020-09-23 #2607 @ethanwharris
@terrytangyuan The authors have now updated the paper to include comparison to other libraries. The new performance comparison and updated examples suggested by @GregaVrbancic are also appreciated. I recommend acceptance :smiley:
