### 2020-09-26 #2607 @jonasrauber
@terrytangyuan Done. Foolbox 3.2.1 is now published on [Zenodo](https://zenodo.org/record/4050932). The DOI is `10.5281/zenodo.4050932` (or `10.5281/zenodo.4050931` for all versions).
