### 2020-09-26 f726939
Fix assertion
