### 2020-09-27 #2607 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
